+++ BlenderAddOn/binjo_addon/binjo_model_bin_collision_seg.py
@@ -280,6 +280,8 @@ class ModelBIN_TriElem:
         self.cull_backface  = True
         self.tex_gen        = False
         self.combiner       = 0
+        self.render_mode    = None
+        self.alpha_compare  = 0
         return
 
     def get_bytes(self):
@@ -306,6 +308,8 @@ class ModelBIN_TriElem:
         self.cull_backface  = True
         self.tex_gen        = False
         self.combiner       = 0
+        self.render_mode    = None
+        self.alpha_compare  = 0
         return
 
     # link the VTX objects in the given VTX-list to the TRI objects in our tri-list via their indices